fraud_analysis_service/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
scikit-learn==1.2.2
pandas==2.0.0
numpy==1.24.2
joblib==1.2.0
python-dotenv==1.0.0